FilterChecker
FilterChecker is a ROBLOX exploit/program that states whether or not a ROBLOX game has the FilteringEnabled property on or off. It can be very useful for when you need to find a game that has the FilteringEnabled property off, so client changes that are made from your exploit will replicate to the server, so everyone can see your thick memes.